first try on block generation
create script for block and for blockGenerator.
the first one is on the block prefabs and will be use by the block to create it or handle event relate to it.
the second one is on the camera and will be used to handle all general function like creating block at start and when the player need new one to be create.
doenst work well at now.